refactor(scorecard): #763 — extract aggregate logic to lib + drop CRAP gate failures
The original `crates/scorecard/src/bin/aggregate.rs` carried `write_scorecard`
(CC 8) and `run` (CC 6) directly in the binary. `cargo llvm-cov nextest
--workspace` doesn't attribute coverage to `src/bin/*.rs` test modules, so
coverage on those functions read as 0.0% even though they had unit tests —
yielding CRAP scores of 72.0 and 42.0 respectively, both well over the
threshold of 15. Same family of issue we fixed for `emit-schema/main` on
PR #764 (commit `5eb347e` style).

Mirror that fix here:

- New `crates/scorecard/src/aggregate.rs` lib module owns all testable
  logic — `parse_cli`, `read_pr_meta`, `validate_against_schema`,
  `ensure_parent_dir` (extracted; reduces `write_scorecard` from CC 8 to
  CC 2), `render_scorecard` (extracted; serialize+validate+pretty-print),
  `write_scorecard`, `run`. Gated under `#[cfg(feature = "cli")]` so the
  lib's mokumo-deps-zero invariant (serde + schemars + serde_json) holds
  under a default `cargo build` — verified via
  `cargo tree -p scorecard --no-default-features`.

- `crates/scorecard/src/bin/aggregate.rs` collapses to a one-line wrapper
  around `scorecard::aggregate::run` (CC 1).

- Tests move with the logic (16 unit tests, +5 over the bin's 11 covering
  the new entry points). All branches of `parse_cli`, `run`, and
  `ensure_parent_dir` are now reachable by `llvm-cov nextest`.

Local crap4rs verification (matches the CI gate):

    Summary: 1431 functions | 0 above threshold (15) | worst: 14.0 | PASS
             avg: 2.0 | median: 1.0 | low: 1305 | acceptable: 72 | moderate: 54

(Previously: 3 above threshold, worst 72.0, FAIL.)

Other gates re-checked locally:
- cargo fmt --all -- --check ✅
- cargo clippy -p scorecard --features cli --all-targets -- -D warnings ✅
- cargo nextest run -p scorecard --features cli → 43/43 ✅
- pnpm --filter='./.github/scripts/scorecard' test → 23/23 ✅
